Fix stuff

See merge request Siegrift/expense-manager-pwa!25

+++ lib/transactions/transaction/transaction.tsx
@@ -19,10 +19,11 @@ import { useDispatch, useSelector } from 'react-redux'
 import { ListChildComponentProps } from 'react-window'
 
 import { Transaction as TransactionState } from '../../addTransaction/state'
+import ConfirmDialog from '../../components/confirmDialog'
 import { useIsBigDevice } from '../../shared/hooks'
 import { formatMoney } from '../../shared/utils'
 import { State } from '../../state'
-import { removeTx, setCursor } from '../actions'
+import { removeTx, setCursor, setConfirmTxDeleteDialogOpen } from '../actions'
 import { applySearchOnTransactions, cursorSel } from '../selectors'
 
 const useStyles = makeStyles((theme: Theme) => ({
@@ -58,6 +59,13 @@ const useStyles = makeStyles((theme: Theme) => ({
   cursor: {
     backgroundColor: 'rgba(0,0,0,0.1)',
   },
+  expenseId: {
+    color: 'gray',
+    fontSize: '0.5em',
+    verticalAlign: 'middle',
+    display: 'inline-block',
+    margin: '0 0 6px 12px',
+  },
 }))
 
 type TransactionContentProps = { tx: TransactionState; bigDevice: boolean }
@@ -66,6 +74,9 @@ const TransactionContent = ({ tx, bigDevice }: TransactionContentProps) => {
   const tags = useSelector((state: State) => state.tags)
   const dispatch = useDispatch()
   const classes = useStyles()
+  const confirmDeleteTxOpen = useSelector(
+    (state: State) => state.transactionList.confirmTxDeleteDialogOpen,
+  )
 
   return (
     <>
@@ -80,6 +91,7 @@ const TransactionContent = ({ tx, bigDevice }: TransactionContentProps) => {
                 tx.amount,
                 tx.currency,
               )}`}
+              <span className={classes.expenseId}>{tx.id.substr(0, 8)}...</span>
             </Typography>
           }
         />
@@ -118,22 +130,42 @@ const TransactionContent = ({ tx, bigDevice }: TransactionContentProps) => {
           {bigDevice && (
             <>
               <Divider orientation="vertical" flexItem style={{ width: 2 }} />
-              <Tooltip title="(E)dit transaction">
-                <Link href={`/transactions/${tx.id}`}>
-                  <IconButton className={classes.iconButton}>
+              <Link href={`/transactions/${tx.id}`}>
+                <Tooltip title="(E)dit transaction">
+                  <IconButton
+                    className={classes.iconButton}
+                    data-cy="edit-icon"
+                  >
                     <EditIcon color="primary" />
                   </IconButton>
-                </Link>
-              </Tooltip>
+                </Tooltip>
+              </Link>
+
               <Tooltip title="(D)elete transaction">
                 <IconButton
                   className={classes.iconButton}
-                  // TODO: confirm dialog
-                  onClick={() => dispatch(removeTx(tx.id))}
+                  onClick={() => dispatch(setConfirmTxDeleteDialogOpen(true))}
                 >
                   <DeleteIcon color="secondary" />
                 </IconButton>
               </Tooltip>
+
+              <ConfirmDialog
+                text={
+                  <>
+                    <p>
+                      Do you really want to remove the following transaction?
+                    </p>
+                    <b>{tx.id}</b>
+                    <p>
+                      <i>This action can't be undone!</i>
+                    </p>
+                  </>
+                }
+                open={confirmDeleteTxOpen}
+                onCancel={() => dispatch(setConfirmTxDeleteDialogOpen(false))}
+                onConfirm={() => dispatch(removeTx(tx.id))}
+              />
             </>
           )}
         </div>
